Move models to its own module
