+++ unitt/src/lib.rs
@@ -1,3 +1,1 @@
-pub mod config;
-pub mod statistics;
-pub mod test;
+pub mod models;
